+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/repository/ContractRepository.java
@@ -1,6 +1,11 @@
 package com.tanaguru.repository;
 
 import com.tanaguru.domain.entity.membership.contract.Contract;
+
+import java.util.Optional;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,4 +14,7 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface ContractRepository extends JpaRepository<Contract, Long> {
+    Page<Contract> findAll(Pageable pageable);
+    Page<Contract> findByNameContainingIgnoreCase(String name, Pageable pageable);
+    Optional<Contract> findById(Long id);
 }