+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/repository/ProjectUserRepository.java
@@ -4,6 +4,8 @@ import com.tanaguru.domain.entity.membership.contract.Contract;
 import com.tanaguru.domain.entity.membership.project.Project;
 import com.tanaguru.domain.entity.membership.project.ProjectAppUser;
 import com.tanaguru.domain.entity.membership.user.User;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -19,7 +21,11 @@ public interface ProjectUserRepository extends JpaRepository<ProjectAppUser, Lon
 
     Collection<ProjectAppUser> findAllByProject_ContractAndContractAppUser_User(Contract contract, User user);
 
+    Page<ProjectAppUser> findAllByProject_ContractAndContractAppUser_User(Contract contract, User user, Pageable pageable);
+
     Collection<ProjectAppUser> findAllByContractAppUser_User(User user);
+
+    Page<ProjectAppUser> findAllByContractAppUser_UserAndProject_NameContainingIgnoreCase(User user, String name, Pageable pageable);
 
     void deleteAllByProject(Project project);
 }