+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/repository/UserRepository.java
@@ -1,6 +1,9 @@
 package com.tanaguru.repository;
 
+import com.tanaguru.domain.entity.membership.contract.Contract;
+import com.tanaguru.domain.entity.membership.project.Project;
 import com.tanaguru.domain.entity.membership.user.User;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,7 +14,8 @@ import java.util.Optional;
  */
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
+    org.springframework.data.domain.Page<User> findAll(Pageable pageable);
+    org.springframework.data.domain.Page<User> findByUsernameContainingIgnoreCaseOrEmailContainingIgnoreCase(String username, String email, Pageable pageable);
     Optional<User> findByUsername(String username);
-
     Optional<User> findByEmail(String email);
 }