+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/service/ProjectService.java
@@ -8,6 +8,8 @@ import com.tanaguru.domain.entity.membership.project.Project;
 import com.tanaguru.domain.entity.membership.project.ProjectAppUser;
 import com.tanaguru.domain.entity.membership.project.ProjectRole;
 import com.tanaguru.domain.entity.membership.user.User;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
 import java.util.Collection;
 import java.util.Optional;
@@ -73,12 +75,21 @@ public interface ProjectService {
      * @return A collection of @see Project
      */
     Collection<Project> findAllByContractAndUser(Contract contract, User user);
+    Page<Project> findPageByContractAndUser(Contract contract, User user, Pageable pageable);
 
     /**
      * @param user     The @see User
      * @return A collection of @see Project
      */
     Collection<Project> findAllByUserMemberOfNotOwner(User user);
+
+    /**
+     * @param user     The @see User
+     * @param pageable page parameter
+     * @param search Search by name string
+     * @return A collection of @see Project
+     */
+    Page<Project> findPageByUserMemberOfNotOwner(User user, String search, Pageable pageable);
 
     /**
      * Find a @see ProjectRole for a given @see EProjectRole